# #12085 Core: Ajax performance (cleanseDomElement): No need to cleanse…