--- a/primefaces/src/main/resources/META-INF/resources/primefaces/core/core.utils.js
+++ b/primefaces/src/main/resources/META-INF/resources/primefaces/core/core.utils.js
@@ -1123,7 +1123,9 @@ if (!PrimeFaces.utils) {
             }
             // Recursively remove events from children elements
             jq.children().each(function() {
-                PrimeFaces.utils.cleanseDomElement($(this), clearData, removeElement);
+                if (this.tagName !== 'OPTION') { //No need to cleanse <option> tag
+                    PrimeFaces.utils.cleanseDomElement($(this), clearData, removeElement);
+                }
             });
 
             // Remove inline event attributes