@cocorossello 2024-06-08
In my test, with the performance measurement script provided in #12085

Without the patch:

```
AJAX updateElement : 104.026123046875 ms
common.js:751 Removed 269 DOM elements
common.js:750 AJAX updateElement : 1.4541015625 ms
common.js:751 Removed 1 DOM elements
common.js:750 AJAX updateElement : 1908.331787109375 ms
common.js:751 Removed 33193 DOM elements
```


With the patch:

```
AJAX updateElement : 118.316162109375 ms
common.js:751 Removed 22 DOM elements
common.js:750 AJAX updateElement : 2.01318359375 ms
common.js:751 Removed 1 DOM elements
common.js:750 AJAX updateElement : 421.157958984375 ms
common.js:751 Removed 2793 DOM elements
```

Numbers are still way worse than @melloware posted in https://github.com/primefaces/primefaces/issues/12085#issuecomment-2155018334, but this is an improvement